+++ vortex-array/src/stats/mod.rs
@@ -23,7 +23,12 @@ use crate::stats::Precision::{Bound, Exact};
 use crate::ArrayData;
 
 pub mod flatbuffers;
+mod ordering;
+mod precision;
 mod statsset;
+
+pub use ordering::{GtOrd, LowerBound, PartialOrder, UpperBound};
+pub use precision::{bound, exact, Precision};
 
 /// Statistics that are used for pruning files (i.e., we want to ensure they are computed when compressing/writing).
 pub const PRUNING_STATS: &[Stat] = &[Stat::Min, Stat::Max, Stat::TrueCount, Stat::NullCount];
@@ -40,223 +45,6 @@ pub const STATS_TO_WRITE: &[Stat] = &[
     Stat::IsStrictSorted,
     Stat::UncompressedSizeInBytes,
 ];
-
-#[derive(Debug, Clone)]
-pub enum Precision<T> {
-    Exact(T),
-    Bound(T),
-}
-
-pub fn exact<S: Into<T>, T>(s: S) -> Precision<T> {
-    Exact(s.into())
-}
-
-pub fn bound<S: Into<T>, T>(s: S) -> Precision<T> {
-    Bound(s.into())
-}
-
-impl<T: Debug> Precision<T> {
-    // This panics if the precision value is not exact
-    pub fn unwrap_exact(self) -> T {
-        match self {
-            Exact(val) => val,
-            _ => vortex_panic!("Expected exact value, got value {:?}", self),
-        }
-    }
-}
-
-impl<T: PartialEq> PartialEq for Precision<T> {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Exact(lhs), Exact(rhs)) | (Bound(lhs), Bound(rhs)) => lhs == rhs,
-            _ => false,
-        }
-    }
-}
-
-impl<T> Precision<Option<T>> {
-    pub fn transpose(self) -> Option<Precision<T>> {
-        match self {
-            Exact(Some(x)) => Some(Exact(x)),
-            Bound(Some(x)) => Some(Bound(x)),
-            _ => None,
-        }
-    }
-}
-
-impl<T> Precision<T> {
-    pub fn ok_exact(self) -> Option<T> {
-        match self {
-            Exact(val) => Some(val),
-            _ => None,
-        }
-    }
-
-    pub fn is_exact(&self) -> bool {
-        matches!(self, Exact(_))
-    }
-
-    pub fn map<U, F: FnOnce(T) -> U>(self, f: F) -> Precision<U> {
-        match self {
-            Exact(value) => Exact(f(value)),
-            Bound(value) => Bound(f(value)),
-        }
-    }
-
-    // Similar to option and then, but if either value is bound, then the whole value is bound.
-    pub fn and_then_prefer_bound<U, F: FnOnce(T) -> Precision<U>>(self, f: F) -> Precision<U> {
-        match self {
-            Exact(value) => f(value),
-            Bound(value) => match f(value) {
-                Exact(value) | Bound(value) => Bound(value),
-            },
-        }
-    }
-
-    fn try_map<U, F: FnOnce(T) -> VortexResult<U>>(self, f: F) -> VortexResult<Precision<U>> {
-        let prec = match self {
-            Exact(value) => Exact(f(value)?),
-            Bound(value) => Bound(f(value)?),
-        };
-        Ok(prec)
-    }
-
-    fn as_ref(&self) -> Precision<&T> {
-        match self {
-            Exact(val) => Exact(val),
-            Bound(val) => Bound(val),
-        }
-    }
-
-    pub fn value(&self) -> &T {
-        match self {
-            Exact(val) | Bound(val) => val,
-        }
-    }
-
-    pub fn into_value(self) -> T {
-        match self {
-            Exact(val) | Bound(val) => val,
-        }
-    }
-}
-
-impl<T: Display> Display for Precision<T> {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Exact(v) => {
-                write!(f, "exact({})", v)
-            }
-            Bound(v) => {
-                write!(f, "bound({})", v)
-            }
-        }
-    }
-}
-
-impl<T: PartialEq> PartialEq<T> for Precision<T> {
-    fn eq(&self, other: &T) -> bool {
-        match self {
-            Exact(v) => v == other,
-            _ => false,
-        }
-    }
-}
-
-impl Precision<ScalarValue> {
-    pub fn into_scalar(self, dtype: DType) -> Precision<Scalar> {
-        self.map(|v| Scalar::new(dtype, v))
-    }
-}
-
-impl Precision<&ScalarValue> {
-    pub fn into_scalar(self, dtype: DType) -> Precision<Scalar> {
-        self.map(|v| Scalar::new(dtype, v.clone()))
-    }
-}
-
-pub trait PartialOrder<T: PartialOrd> {
-    fn ordered(lhs: &T, other: &T) -> Option<bool>;
-
-    fn lift(value: Precision<T>) -> Self;
-
-    fn into_value(self) -> Precision<T>;
-}
-
-pub struct LowerBound<T>(Precision<T>);
-
-impl<T> LowerBound<T> {
-    pub fn is_exact(&self) -> bool {
-        self.0.is_exact()
-    }
-}
-
-impl<T: PartialOrd> LowerBound<T> {
-    pub fn le(&self, value: &LowerBound<T>) -> Option<bool> {
-        Some(match self.0.value().partial_cmp(value.0.value())? {
-            Ordering::Less => true,
-            Ordering::Equal => {
-                // for a fixed value v. exact(v) <= bound(v) is true
-                matches!((&self.0, &value.0), (Exact(_), _) | (Bound(_), Bound(_)))
-            }
-            Ordering::Greater => false,
-        })
-    }
-}
-
-impl<T: PartialOrd> PartialOrder<T> for LowerBound<T> {
-    fn ordered(lhs: &T, rhs: &T) -> Option<bool> {
-        PartialOrd::partial_cmp(lhs, rhs).map(|o| o != Ordering::Greater)
-    }
-
-    fn lift(value: Precision<T>) -> Self {
-        Self(value)
-    }
-
-    fn into_value(self) -> Precision<T> {
-        self.0
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct UpperBound<T>(Precision<T>);
-
-pub trait GtOrd<Rhs: ?Sized = Self> {
-    fn ge(&self, other: &Rhs) -> Option<bool>;
-}
-
-impl<T: PartialOrd> GtOrd for UpperBound<T> {
-    fn ge(&self, other: &Self) -> Option<bool> {
-        Some(match self.0.value().partial_cmp(other.0.value())? {
-            Ordering::Less => false,
-            Ordering::Equal => {
-                // for a fixed value v. exact(v) >= bound(v) is true
-                matches!((&self.0, &other.0), (Exact(_), _) | (Bound(_), Bound(_)))
-            }
-            Ordering::Greater => true,
-        })
-    }
-}
-
-impl<T: PartialOrd> GtOrd<T> for UpperBound<T> {
-    fn ge(&self, other: &T) -> Option<bool> {
-        Some(self.0.value() >= other)
-    }
-}
-
-impl<T: PartialOrd> PartialOrder<T> for UpperBound<T> {
-    fn ordered(lhs: &T, rhs: &T) -> Option<bool> {
-        PartialOrd::partial_cmp(lhs, rhs).map(|o| o != Ordering::Less)
-    }
-
-    fn lift(value: Precision<T>) -> Self {
-        Self(value)
-    }
-
-    fn into_value(self) -> Precision<T> {
-        self.0
-    }
-}
 
 pub trait StatOrder<T: PartialOrd> {
     type BoundDirection: PartialOrder<T>;
